NJTA-10-added-tests-with-coverage

+++ src/main/java/net/njit/ms/cs/model/dto/DepartmentDto.java
@@ -15,5 +15,5 @@ public class DepartmentDto {
 
     private String chairSsn;
 
-    private Long buildingNumber;
+    private Integer buildingNumber;
 }
